backend/requirements.txt
quart
quart-cors
google-genai
python-dotenv
hypercorn
